Backend/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
keras==2.14.0
kiwisolver==1.4.8
libclang==18.1.1
Mako==1.3.9
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
openai==1.72.0
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
propcache==0.3.1
protobuf==4.25.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.28.2
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sympy==1.13.1
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
threadpoolctl==3.6.0
torch==2.6.0
tqdm==4.67.1
twilio==9.5.2
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==1.26.20
Werkzeug==3.1.3
wrapt==1.14.1
WTForms==3.0.1
yarl==1.19.0
